requirements.txt
flask>=2.0.0
pillow>=9.0.0
numpy>=1.22.0
opencv-python>=4.7.0
ultralytics>=8.0.0
torch>=2.0.0
python-dotenv==1.0.0
requests>=2.28.0 